Fix current version resolving for clients
