+++ src/Paysera/Bundle/ClientReleaseBundle/Service/ClientDefinitionBuilder/JavascriptClientDefinitionBuilder.php
@@ -5,14 +5,17 @@ namespace Paysera\Bundle\ClientReleaseBundle\Service\ClientDefinitionBuilder;
 
 use Paysera\Bundle\ClientReleaseBundle\Entity\ClientDefinition;
 use Paysera\Bundle\ClientReleaseBundle\Entity\JavascriptClientDefinition;
+use Paysera\Bundle\ClientReleaseBundle\Service\VersionResolver\VersionResolverInterface;
 
 class JavascriptClientDefinitionBuilder implements ClientDefinitionBuilderInterface
 {
     private $baseDefinitionBuilder;
+    private $versionResolver;
 
-    public function __construct(BaseDefinitionBuilder $baseDefinitionBuilder)
+    public function __construct(BaseDefinitionBuilder $baseDefinitionBuilder, VersionResolverInterface $versionResolver)
     {
         $this->baseDefinitionBuilder = $baseDefinitionBuilder;
+        $this->versionResolver = $versionResolver;
     }
 
     public function buildClientDefinition(array $data): ClientDefinition
@@ -23,6 +26,8 @@ class JavascriptClientDefinitionBuilder implements ClientDefinitionBuilderInterf
         if (isset($data['client_name'])) {
             $definition->setClientName($data['client_name']);
         }
+
+        $definition->setVersionResolver($this->versionResolver);
 
         return $definition;
     }
