+++ src/Paysera/Bundle/ClientReleaseBundle/Service/ClientDefinitionBuilder/PhpClientDefinitionBuilder.php
@@ -5,14 +5,17 @@ namespace Paysera\Bundle\ClientReleaseBundle\Service\ClientDefinitionBuilder;
 
 use Paysera\Bundle\ClientReleaseBundle\Entity\ClientDefinition;
 use Paysera\Bundle\ClientReleaseBundle\Entity\PhpClientDefinition;
+use Paysera\Bundle\ClientReleaseBundle\Service\VersionResolver\VersionResolverInterface;
 
 class PhpClientDefinitionBuilder implements ClientDefinitionBuilderInterface
 {
     private $baseDefinitionBuilder;
+    private $versionResolver;
 
-    public function __construct(BaseDefinitionBuilder $baseDefinitionBuilder)
+    public function __construct(BaseDefinitionBuilder $baseDefinitionBuilder, VersionResolverInterface $versionResolver)
     {
         $this->baseDefinitionBuilder = $baseDefinitionBuilder;
+        $this->versionResolver = $versionResolver;
     }
 
     public function buildClientDefinition(array $data): ClientDefinition
@@ -23,6 +26,8 @@ class PhpClientDefinitionBuilder implements ClientDefinitionBuilderInterface
         if (isset($data['namespace'])) {
             $definition->setNamespace($data['namespace']);
         }
+
+        $definition->setVersionResolver($this->versionResolver);
 
         return $definition;
     }
