+++ src/Paysera/Bundle/ClientReleaseBundle/Service/VersionResolver/VersionResolverInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paysera\Bundle\ClientReleaseBundle\Service\VersionResolver;
+
+use Paysera\Bundle\ClientReleaseBundle\Entity\ReleaseStepData;
+
+interface VersionResolverInterface
+{
+    const VERSION_DEFAULT = '0.0.0';
+
+    public function resolveCurrentVersion(ReleaseStepData $releaseStepData): string;
+}
